+++ code/script.ts
@@ -1,12 +1,9 @@
 /*
 geändert bis zur nächsten Version
-- Eingabefeld wird geleert, wenn es angeklickt wird
-- Sicherheitskopien werden nur in einem Verzeichnis gespeichert
-- Verzeichnis mit Sicherheitskopien schnell-tagger_sec wird nicht mehr angezeigt
 
 */
 
-console.log('Schnell-Tagger Version 0.2; AGPL 3: https://www.gnu.org/licenses/agpl-3.0.de.html, Autor und Credit: Wolf Hosbach, http://www.wolf-hosbach.de, https://github.com/wolfhos/schnell-tagger');
+console.log('Schnell-Tagger Version 0.3; AGPL 3: https://www.gnu.org/licenses/agpl-3.0.de.html, Autor und Credit: Wolf Hosbach, http://www.wolf-hosbach.de, https://github.com/wolfhos/schnell-tagger');
 
 
 //Konfiguration hier ändern