+++ CHANGELOG.md
@@ -1,5 +1,13 @@
 # Change Log
 
+## v0.3
+
+* Löschen von Stichwörtern
+* Eingabefeld wird geleert, wenn es angeklickt wird
+* Sicherheitskopien werden nur in einem Verzeichnis gespeichert
+* Verzeichnis mit Sicherheitskopien schnell-tagger_sec wird nicht mehr angezeigt
+
+
 ## v0.2
 * INI-Datei für alle IPTC-Felder ergänzt. Hier kann auch ein Urheber eingestellt werden. UTF-8 ist voreingestellt. Das beseitigt den Bug, dass alle vorhandenen Felder überschrieben wurden.
 * Eingabefeld wird geleert, wenn es angeklickt wird